Move cookie to Common/Core as it has nothing to do with that type of header
