+++ src/Common/Core/Cookie.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace Common\Core\Header;
+namespace Common\Core;
 
-use Common\Core\Model;
 use Symfony\Component\HttpFoundation\Cookie as SymfonyCookie;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\ResponseHeaderBag;
